+++ templates/hexarch/App/test/Fixture/ApplicationFixtureCollection.cs
@@ -1,0 +1,6 @@
+using Arbeidstilsynet.HexagonalArchitectureTemplateDocker.App.Test.Fixture;
+
+namespace Arbeidstilsynet.HexagonalArchitectureTemplateDocker.App.Test;
+
+[CollectionDefinition("Application Fixture Collection")]
+public class ApplicationFixtureCollection : ICollectionFixture<ApplicationFixture> { }