+++ templates/hexarch/Domain/Logic/src/SakService.cs
@@ -1,0 +1,60 @@
+using Arbeidstilsynet.HexagonalArchitectureTemplateDocker.Domain.Data;
+using Arbeidstilsynet.HexagonalArchitectureTemplateDocker.Domain.Logic.DependencyInjection;
+using Arbeidstilsynet.HexagonalArchitectureTemplateDocker.Domain.Ports.App;
+using Arbeidstilsynet.HexagonalArchitectureTemplateDocker.Domain.Ports.App.Requests;
+using Arbeidstilsynet.HexagonalArchitectureTemplateDocker.Domain.Ports.Infrastructure;
+using Microsoft.Extensions.Options;
+
+namespace Arbeidstilsynet.HexagonalArchitectureTemplateDocker.Domain.Logic;
+
+internal class SakService(ISakRepository sakRepository, IOptions<DomainConfiguration> domainConfig)
+    : ISakService
+{
+    public async Task<Sak> CreateNewSak(CreateSakDto sakDto)
+    {
+        using var activity = Tracer.Source.StartActivity();
+
+        var createdAt = DateTime.UtcNow;
+        var deadline = createdAt + TimeSpan.FromDays(domainConfig.Value.SakDeadlineDays);
+
+        var sak = new Sak
+        {
+            Id = Guid.NewGuid(),
+            CreatedAt = createdAt,
+            LastUpdated = createdAt,
+            Organisasjonsnummer = sakDto.Organisasjonsnummer,
+            Status = SakStatus.New,
+            Deadline = deadline,
+        };
+
+        return await sakRepository.PersistSak(sak);
+    }
+
+    public async Task<IEnumerable<Sak>> GetAllSaker()
+    {
+        return await sakRepository.GetSaker();
+    }
+
+    public async Task<Sak> ArchiveSak(Guid sakId)
+    {
+        return await sakRepository.UpdateSakStatus(sakId, SakStatus.Archived)
+            ?? throw new SakNotFoundException(sakId);
+    }
+
+    public async Task<Sak> EndSak(Guid sakId)
+    {
+        return await sakRepository.UpdateSakStatus(sakId, SakStatus.Done)
+            ?? throw new SakNotFoundException(sakId);
+    }
+
+    public async Task<Sak> StartSak(Guid sakId)
+    {
+        return await sakRepository.UpdateSakStatus(sakId, SakStatus.InProgress)
+            ?? throw new SakNotFoundException(sakId);
+    }
+
+    public async Task<Sak> GetSakById(Guid sakId)
+    {
+        return await sakRepository.GetSak(sakId) ?? throw new SakNotFoundException(sakId);
+    }
+}