+++ templates/hexarch/Domain/Ports/App/AssemblyInfo.cs
@@ -1,0 +1,6 @@
+﻿namespace Arbeidstilsynet.HexagonalArchitectureTemplateDocker.Domain.Ports.App;
+
+/// <summary>
+/// Assembly information marker interface.
+/// </summary>
+public interface IAssemblyInfo { }