+++ templates/hexarch/Infrastructure/Infrastructure.Adapters/Db/Model/BaseEntity.cs
@@ -1,7 +1,0 @@
-namespace Arbeidstilsynet.HexagonalArchitectureTemplateDocker.Infrastructure.Adapters.Db.Model;
-
-internal abstract class BaseEntity
-{
-    public DateTime CreatedAt { get; set; }
-    public DateTime UpdatedAt { get; set; }
-}