+++ templates/hexarch/Infrastructure/Infrastructure.Ports/AssemblyInfo.cs
@@ -1,3 +1,0 @@
-﻿namespace Arbeidstilsynet.HexagonalArchitectureTemplateDocker.Infrastructure.Ports;
-
-public interface IAssemblyInfo { }